+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## 2.1.0 - 2016-10-13
 ### Added
 - `lookup()` method to search the store by IDs
+- 
 
 ## 2.0.1 - 2016-10-13
 ### Fixed